backend/requirements.txt
asgiref==3.11.1
boto3==1.42.97
botocore==1.42.97
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==47.0.0
Django==4.2.30
django-cors-headers==4.9.0
django-environ==0.13.0
django-ratelimit==4.1.0
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
google-auth==2.50.0
gunicorn==23.0.0
idna==3.13
jmespath==1.1.0
pillow==11.3.0
psycopg2-binary==2.9.12
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.12.1
python-dateutil==2.9.0.post0
razorpay==2.0.1
requests==2.32.5
s3transfer==0.16.1
six==1.17.0
sqlparse==0.5.5
stripe==15.1.0
typing_extensions==4.15.0
urllib3==1.26.20
